+++ doc/changelog.rst
@@ -111,6 +111,7 @@ Addons
 Front
 ~~~~~
 
+- Add login check for inactive contacts
 - Remove "Ordering for company" from checkout if logged in
 - Allow user to link company contact from account page
 - Log-in as company if user is a member of a company